support "include data from all cycles" option in backend
