+++ server/system/apiRouter.js
@@ -10,6 +10,7 @@ import * as geoApi from '@server/modules/geo/api/geoApi'
 import * as collectImportApi from '@server/modules/collectImport/api/collectImportApi'
 import * as arenaImportApi from '@server/modules/arenaImport/api/arenaImportApi'
 import * as recordApi from '@server/modules/record/api/recordApi'
+import * as dataExportApi from '@server/modules/dataExport/api/dataExportApi'
 import * as surveyRdbApi from '@server/modules/surveyRdb/api/surveyRdbApi'
 import * as reportingApi from '@server/modules/reporting/api'
 import * as analysisApi from '@server/modules/analysis/api'
@@ -35,6 +36,7 @@ collectImportApi.init(router)
 arenaImportApi.init(router)
 
 recordApi.init(router)
+dataExportApi.init(router)
 
 surveyRdbApi.init(router)
 reportingApi.init(router)
